ANDROID: drm/virtio: track {fence_id, fence_info} in virtio_gpu_fence
Each fence should be associated with a [fence ID, fence_context,
seqno].  The seqno number is just the fence id.

To get the fence context, we add the fence_context_idx to the
3D context's base_fence_ctx.  The fence_context_idx is between 0 and
31, inclusive.

Each 3D context will have it's own base_fence_ctx. The fence_context_idx
will be emitted to host userspace, when emit_fence_info is true.

Signed-off-by: Gurchetan Singh <gurchetansingh@chromium.org>

BUG=b:173630595
TEST=boot VM

Change-Id: I78342ae267adf9050305716076c2bc5b357d11c1