Fix issue where wlan_simulation and goldfish drivers weren't built
When doing android-mainline, I was wondering why no files in
modules_staging_dir...extra were created, leading back to bunch of
missing modules in extra for cf/gf in android12-5.4 as well.

Can't use just any CONFIG_* in defconfigs here to select stuff to
build, has to be from existing set. So wlan_simulation is now built when
CONFIG_VIRT_WIFI. There seems to be a CONFIG_GOLDFISH in common/ for the
goldfish drivers, but it depends on X86_CONFIG_GOLDFISH which doesn't
seem to be available on aarch64, so it uses a plain environment variable
instead.

Bug: 166154615
Signed-off-by: Lingfeng Yang <lfy@google.com>
Change-Id: Iaa07eed14018329a3117d0ed46414e717c637bed