ANDROID: create a configuration for the ARM Fixed Virtual Platform
This config is based on gki_defconfig but includes drivers required to
support the FVP's emulated hardware.

The enabled drivers are sufficient to boot the non-GUI fvp_mini lunch
target, but not the main fvp target which relies on FBDEV drivers. A
future change will enable the appropriate pl111 DRM drivers once fvp
can be switched over to DRM.

Bug: 142352330
Change-Id: Ifedce69a0250403775344d67499fe3888899c3f6
Signed-off-by: Peter Collingbourne <pcc@google.com>