ANDROID: add crashdump.<arch> configs
These are new configs for the kernel to use for kdump in Microdroid. The
kernel is loaded as a panic kernel (KEXEC_FILE_ON_CRASH) there. Since
the only task there is dumping /proc/vmcore to the host-side via a
virtio-console device, the config is very minimal.

Bug: 237380607
Test: follow the steps shown at
https: //android-review.googlesource.com/c/platform/packages/modules/Virtualization/+/2147677
Signed-off-by: Jiyong Park <jiyong@google.com>
Change-Id: Ia8a45472ed8c84a3497f80b4ed1f1cbdedfe0992