Fix trivial typo in defconfig_test.h
CFG80211 was changed from build-in to a vendor module in commit 3a951e6
("Add CFG80211 and MAC80211 as modules"). The error message was updated
to reflect this. However, a minor typo was made so fix it.

Signed-off-by: Carlos Llamas <cmllamas@google.com>
Change-Id: Ia97a9723d35f092f3a4ae2779504fa0cc1c5eef3
(cherry picked from commit 827e53c41c500a005d208bbadc3a03b34ecc36eb)