kleaf: enable DDK
Use the DDK (Driver development kit) to build the kernel module.

To achieve that, to avoid the use of an unsafe header
drivers/gpu/drm/virtio/virtgpu_trace.h, it is copied to
this folder.

Also filter out the proper sources for building in-tree modules, so
that changes in sources for external modules does not affect
the kernel_build.

In this change, a macro called virtual_device_externel_modules
is also introduced. Though we recommend DAMP (descriptive and
meaningful phrases) vs DRY (don't repeat yourself) for BUILD
files, the macro is an adequate abstraction layer because, without
ddk_modules, all these virtual device targets would share the same
Kbuild files as well, meaning that they'll have similar
configurations.

Bug: 254735056
Test: bazel build //common-modules/virtual-device:virtual_device_x86_64_modules_install

Change-Id: I66a30cc4b9ab66e55daba048023528bd0a5dd46e
Signed-off-by: Yifan Hong <elsk@google.com>