CHROMIUM: Add dev_sw_sync driver
ARC devices starting with Android version R or greater require that the
debugfs system is not mounted or used. Therefore, in order to support
the use of sw_sync fences by the ARC surfaceflinger, the arcvm kernel
needs to implement a new driver that implements sw_sync using a kernel
device instead of a debugfs file to interact with it from user space.

Here the upstream kernel implementation is kept in tact, but a new
staging driver is added that uses /dev/sw_sync to call the sw_sync APIs.
The new driver is a port of the arcvm implementation as an out of tree
module.

This has been done in this CL as well so that dev-sw-sync driver is used by
default for the x86_64_arcvm kernel targets (ARC targets).

BUG: 259180445

Signed-off-by: yulilin <yulilin@google.com>
Signed-off-by: Stan Rokita <srok@google.com>

Change-Id: I37809dd5f852912480e28de93266297b12940a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3301104
Reviewed-by: Lepton Wu <lepton@chromium.org>
Commit-Queue: Stan Rokita <srok@google.com>
Tested-by: Stan Rokita <srok@google.com>

dev-sw-sync as a module

Change-Id: I825981c28f369fb6ad4d1dd681efd65b83295dfb