kleaf: Add arm and i686 target.
Unlike the arm64 / x86_64 targets, the arm and i686
target does not use mixed build. This is because
we don't have an arm / i686 GKI target. Hence, this needs
to explicitly declare the list of GKI modules in
module_outs.

In addition, the default dist directory is set to
out/{name}/dist to avoid the legacy behavior of reading
BRANCH.

Test: TH
Test: manually build it with
  bazel build //common-modules/virtual-device:virtual_device_arm_dist
  bazel build //common-modules/virtual-device:virtual_device_i686_dist
Bug: 258308697
Bug: 231753340
Change-Id: Id35048df354efb7c25b15224cad4ebe7a3a1a0a9