Crashdump kernel is built like the microdroid kernel
Previously, the build rules for crashdump kernel depended on the generic
kernel. As a result, its size was considerable larger than was expected
(3MB v.s. 34MB).

This change fixes the issue by dropping the dependencies to the generic
kernel. Now the build rules for crashdump kernel is very similar to
those of the microdroid kernel except the config file names.

Bug: 259008494
Test: tools/bazel run --lto=thin
//common-modules/virtual-device:crashdump_aarch64_dist
du -sh out/android14-5.15/dist/Image
is 3.9MB (was 34MB)

Change-Id: I3f65b9fbcdb9591c2da6062f8baff2bc7b99545e