kleaf: virtual_device: Replace COMMON_GKI_MODULES_LIST with get_gki_modules_list()
get_gki_modules_list(arch) is introduced as a replacement to
COMMON_GKI_MODULES_LIST to get list of common GKI modules
for the architecture as this list can be different for
different archintecture.

get_gki_modules_list() only supports arm64, riscv64 and
x86_64. Use arm64 for arm and x86_64 for i386 as the
list of modules should be same for these architectures.

Bug: 293529933
Test: TH
Test: bazel run //common-modules/virtual-device:virtual_device_arm_dist
Change-Id: I55a9621d7348635ec2914551667ae421936cd143
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>