ANDROID: drivers: hv: dxgkrnl: Implement support for mapping guest pages on the host.
Prior to this change GPADLs were user to represent system memory buffers
to the host. GPADL can represent a buffer with max of ~32MB and there is
also a limit of total size of GPADLs in a VM.

The change makes use of the API on host, which maps guest pages. This API
does not have GPADL limitations.

Signed-off-by: Iouri Tarassov <iourit@linux.microsoft.com>

(am from
https://github.com/microsoft/WSL2-Linux-Kernel/commit/17d7a3fab59b614bfa3429a6e7bc91f07ed4592a)

BUG=308813188
TEST=ran android VM at end of chain

Change-Id: I4f264af50172a1bd9e06d9d59bfe13bab3309afa
Signed-off-by: Colin Downs-Razouk <colindr@google.com>
Signed-off-by: Kriti <kritisn@google.com>