ANDROID: drivers: hv: dxgkrnl: Implement DXGSYNCFILE
Implement the LX_DXCREATESYNCFILE IOCTL (D3DKMTCreateSyncFile).

dxgsyncfile is built on top of the Linux sync_file object and
provides a way for the user mode to synchronize with the execution
of the device DMA packets.

The IOCTL creates a dxgsyncfile object for the given GPU synchronization
object and a fence value. A sync_object file descriptor is returned to
the caller. The caller could wait for the object by using poll().
When the GPU synchronization object is signaled on the host, the host
sends a message to the virtual machine and the sync_file object is
signaled.

Signed-off-by: Iouri Tarassov <iourit@linux.microsoft.com>

(am from
https://github.com/microsoft/WSL2-Linux-Kernel/commit/026b351921973de8653afe8ecdaf479f7c8a6b94)

BUG=308813188
TEST=ran android VM at end of chain

Change-Id: I7b6245517ccc7cb19532b1cbfeb362eaf82d844b
Signed-off-by: Colin Downs-Razouk <colindr@google.com>