ANDROID: virtio-dxgkrnl: present virtual
This present api allows the guest HWC to pass an acquire and release
semaphore handle, as well as a exported memory handle to the host for
presentation. The handles should be the same nthandles that are returned
from ShareObjects. If either the acquire or release semaphore handles is
null then it is ignored. The driver can also pass private data required
for presentation.

Bug: 278593677
Test: built kernel
Change-Id: I9a5e8d9d7fd8d8f88b5dc5cf1850c746e37ed6b6
Signed-off-by: Colin Downs-Razouk <colindr@google.com>