FUSE: set up module to implement tmpfiles and other vendor hooks
The FUSE tmpfile implementation in Linux upstream cannot be backported
to android13-5.15 because it breaks the ABI. Set up FUSE vendor module
to provide provide a non-atomic implementation of tmpfiles that can
connect to the vendor hooks in the common kernel. Allow handlers for
other vendor hooks to be added easily.

Vendor hooks for tmpfiles are added to the android13-5.15 common kernel
in aosp/2956757.

BUG=323975715
BUG=318581825
TEST=arc.Boot.vm,
     cts -m CtsBionicTestCases -t dlext#dlopen_ext_use_o_tmpfile_fd,
     cts -m CtsPerfettoTestCases -t PerfettoCmdlineTest#NoDataNoFileWithoutTrigger

Change-Id: I069f7943670d2c4891106e78cb5b1fac8cb4e1d8
Signed-off-by: Tiffany Yang <ynaffit@google.com>