linux_distro.fragment: tweaking CONFIG to be more close to Debian/Ubuntu
This commit fixes some issues.

CONFIG_FHANDLE is needed by systemd. Otherwise some service failed randomly.
It fixes the issue of rootfs not remounted as read-write sometimes.

CONFIG_MODPROBE_PATH is needed by cuttlefish. Because cuttlefish uses
iptables, and we need this PATH to be set to load iptables related modules.

And also fixes the following issues that we don't know which CONFIG really
fixes the issue. The first is tty1 was not working on some of the servers.
Probably nVIDIA video card servers was not working. AMDGPU servers works
fine always. Now all of the server works. The second issue is the performance
to run cuttlefish which was much slower than Ubuntu. After applying this
commit the performance becomes normal.

Change-Id: I15c97a913b3c790e38e32632eabbc2791790e7bc
Signed-off-by: Ying-Chun Liu <liupa@google.com>