i686 use defconfig fragments.
This allows us to check that the configs in the
corresponding fragments are actually enabled properly.

This change also drops virtio_pmem.ko
and nd_virtio.ko because
it depends on LIBNVDIMM which is only enabled
on 64-bit. PHYS_ADDR_T_64BIT was previously
incorrectly enabled because we used merge_config
with the incorrect gki_defconfig as base.

Other architectures will be fixed in other CLs. In
case of any issues, this allows for a smaller
revert.

Test: TH
Change-Id: Ia6c8c40eeb5e8e38434eff347bb87b1bdb45b21b