aarch64 16k use defconfig fragments.
This allows us to check that the configs in the
corresponding fragments are actually enabled properly.

Because the build config is no different from
the 4k build config after dropping PRE_DEFCONFIG_CMDS,
delete build.config.virtual_device.aarch64.16k.

Other architectures will be fixed in other CLs. In
case of any issues, this allows for a smaller
revert.

Test: TH
Change-Id: I89e27910e2f75bac4426608bd4dbdc78165f200f