aarch64 16k use defconfig fragments.
This allows us to check that the configs in the
corresponding fragments are actually enabled properly.

Because the build config is no different from
the 4k build config after dropping PRE_DEFCONFIG_CMDS,
delete build.config.virtual_device.aarch64.16k.

Other architectures will be fixed in other CLs. In
case of any issues, this allows for a smaller
revert.

Due to b/345544754, CONFIG_ARCH_MMAP_RND_COMPAT_BITS
needs to be explicitly set.

This change also changes `16k_gki.fragment` to
the native kernel_build.page_size=16k. A notable
effect of this change is that CONFIG_LOCALVERSION
changes from 4k to empty, which is expected
because 16k_gki.fragment is outdated.

Test: TH
Bug: 345544754
Change-Id: I89e27910e2f75bac4426608bd4dbdc78165f200f