kleaf: add test for textual_hdrs exporting include properly.
Previously, if a ddk_headers target is provided in textual_hdrs, the
ddk_headers.includes are only exported to dependent modules of this
ddk_module, but not this ddk_module. Fix that.

This is so that textual_hdrs is treated the
same way as hdrs.

Bug: 355075293
Change-Id: I4a759ef8651e7605960ee7fef6520bf1fab1415d
(cherry picked from commit 4fdeac11867adf823b127537ff5b39d9ff3eb36f)