Fix dtb Makefile for rockpi.
Fixes the issue of Makefile.dtbs not being included when dtb-, dtb-y
etc. is not set, introduced by e7e2941300d2
("kbuild: split device tree build rules into scripts/Makefile.dtbs").

Implementation note:

I intentionally use dtb-, not dtb-y, so that rockpi4 must explicitly
put the dtb name in make_goals to build it. This is so that the Makefile
has minimal effect on other targets (if there are any).

Bug: 371628764
Change-Id: Ifd117183d1bf562dc8afccc93879646aa26d8a53