ANDROID: virtio-dxgkrnl: when mapping memory grab the write lock and set VM_DONTCOPY
We refcount the mapped memory in lock2/unlock2. However, that doesn't
work well with fork - when unlock2 in one process destroys that memory
mapping, there is another page table entry left in the other process
still pointing to that same GPA. And this can confuse the PAT tracing,
and cause memory mapping on the same GPA to fail under certain
circumstances.

Test: presubmit
Bug: b/364907812
Change-Id: Idf74ed2d9c41683bb6fb821d26f698ab79a7f8bf
Signed-off-by: Kaiyi Li <kaiyili@google.com>